requirements.txt
streamlit
langchain_groq
langchain_community
langchain
python-dotenv
sentence-transformers
langchain_huggingface
bs4
faiss-cpu